Keep plugin detail rendering out of the startup bundle
