+++ apps/app/src/components/plugin/plugin-detail-navigation.tsx
@@ -13,42 +13,11 @@ import {
 import { PluginIcon } from "./PluginIcon";
 import type { ThreadSecondaryPanelProps } from "@/components/secondary-panel/ThreadSecondaryPanel";
 import { SecondaryPanelContentSkeleton } from "@/components/secondary-panel/lazySecondaryPanelComponents";
-
-interface PluginDetailDestination {
-  pluginId: string;
-  title: string;
-}
-
-type PluginDetailOpener = (destination: PluginDetailDestination) => boolean;
-
-const focusedOpeners = new Map<symbol, PluginDetailOpener>();
-
-export function openPluginDetailsInWorkspace(
-  destination: PluginDetailDestination,
-): boolean {
-  for (const open of [...focusedOpeners.values()].reverse()) {
-    if (open(destination)) return true;
-  }
-  return false;
-}
-
-export function usePublishPluginDetailOpener(
-  open: PluginDetailOpener,
-  isActive: boolean,
-): void {
-  const openRef = useRef(open);
-  useLayoutEffect(() => {
-    openRef.current = open;
-  }, [open]);
-  useLayoutEffect(() => {
-    if (!isActive) return;
-    const token = Symbol("plugin-detail-opener");
-    focusedOpeners.set(token, (destination) => openRef.current(destination));
-    return () => {
-      focusedOpeners.delete(token);
-    };
-  }, [isActive]);
-}
+import {
+  usePublishPluginDetailOpener,
+  type PluginDetailDestination,
+  type PluginDetailOpener,
+} from "./plugin-detail-opener";
 
 const LazyPluginDetailPaneView = lazy(() =>
   import("@/views/ToolsView").then(({ PluginDetailPaneView }) => ({
